fix(backend): check passkey length before deletion (#185)

+++ apps/backend/src/api/core/helpers/webauthn/registration/index.ts
@@ -106,7 +106,8 @@ export default class WebAuthnRegistration extends SingletonBase implements IWebA
       throw Error(`${this.constructor.name} | complete | Invalid challenge type for ${userIdentifier}`)
 
     // Delete existing passkeys
-    await this.passkeyRepository.deletePasskeys(user.passkeys.map(passkey => passkey.credentialId))
+    if (user.passkeys.length)
+      await this.passkeyRepository.deletePasskeys(user.passkeys.map(passkey => passkey.credentialId))
     // Create new passkey
     const passkey = await this.passkeyRepository.createPasskey(
       {
